add reverse option in MultiGetOptions to support search in reverse direction for multiGet
Summary: Ref T10566

Test Plan: N/A

Reviewers: sunweijie, cailiuyang

Reviewed By: sunweijie

Subscribers: #pegasus

Maniphest Tasks: T10566

Differential Revision: https://phabricator.d.xiaomi.net/D87870

+++ idl/recompile_thrift.sh
@@ -1,14 +1,22 @@
 #!/bin/bash
 
 # method to get thrift 0.11.0:
-# 1. get the source code from http://mirrors.tuna.tsinghua.edu.cn/apache/thrift/0.11.0/thrift-0.11.0.tar.gz
-# 2. tar xvf thrift-0.11.0.tar.gz & cd thrift-0.11.0
-# 3. ./configure --with-csharp=no --with-java=no --with-python=no --with-erlang=no --with-perl=no --with-php=no --with-ruby=no --with-haskell=no --with-php_extension=no --with-rs=no
-# 4. make && sudo make install
+#   wget http://mirrors.tuna.tsinghua.edu.cn/apache/thrift/0.11.0/thrift-0.11.0.tar.gz
+#   tar xvf thrift-0.11.0.tar.gz
+#   cd thrift-0.11.0
+#   ./configure --with-csharp=no --with-java=no --with-python=no --with-erlang=no --with-perl=no \
+#       --with-php=no --with-ruby=no --with-haskell=no --with-php_extension=no --with-rs=no --with-go=no
+#   make
+#   sudo make install
 # 
-# Note: by default the thrift will be installed in /usr/local. If you'd like to install it somewhere else, you'd better with --prefix=... in step 3
+# Note: by default the thrift will be installed in /usr/local. If you'd like to install it somewhere else,
+# you'd better provide install path with option `--prefix=xxx'.
  
 thrift=thrift
+if ! $thrift -version | grep "0.11.0" ; then
+    echo "ERROR: thrift version should be 0.11.0"
+    exit 1
+fi
 
 TMP_DIR=./gen-java
 rm -rf $TMP_DIR
